Check the m_process object before calling the terminate() method. In some situations (when the helper app isn't found), the m_process object is not even created and this will lead to a crash
BUG: 192798

svn path=/trunk/KDE/kdeutils/ark/; revision=968443

+++ kerfuffle/cliinterface.cpp
@@ -416,7 +416,8 @@ namespace Kerfuffle
 		//KProcess *p = m_process;
 		//m_process = NULL;
 		//if (p) p->terminate();
-		m_process->terminate();
+		if (m_process)
+			m_process->terminate();
 		finished(false);
 	}
